refactor(tag): make display tag route the  main route
The old tags/$tag route is moved to tags/single/$tag.
tags/single is used by the questions route and so could be removed if questions route is removed.
