+++ app/routes/questions.$questionId.tsx
@@ -5,7 +5,7 @@ import AutoHeight from 'react-auto-height'
 import type {Question, Glossary, PageId, Banner as BannerType, Tag} from '~/server-utils/stampy'
 import type useQuestionStateInUrl from '~/hooks/useQuestionStateInUrl'
 import {Edit, Link as LinkIcon} from '~/components/icons-generated'
-import {Tags} from '~/routes/tags.$tag'
+import {Tags} from '~/routes/tags.single.$tag'
 import CopyLink from '~/components/copyLink'
 import {Action, ActionType} from '~/routes/questions.actions'
 import {reloadInBackgroundIfNeeded} from '~/server-utils/kv-cache'
